refactor: merged PostPermissionEvaluator and PostPermissionPolicies into PostPermissionResolver
